+++ engine/samples/voxels/main.cpp
@@ -1,6 +1,11 @@
+#include <cubos/engine/assets/plugin.hpp>
+#include <cubos/engine/render/camera/draws_to.hpp>
+#include <cubos/engine/render/camera/perspective_camera.hpp>
+#include <cubos/engine/render/defaults/plugin.hpp>
+#include <cubos/engine/render/defaults/target.hpp>
 #include <cubos/engine/render/lights/directional.hpp>
-#include <cubos/engine/renderer/plugin.hpp>
-#include <cubos/engine/screen_picker/plugin.hpp>
+#include <cubos/engine/render/voxels/grid.hpp>
+#include <cubos/engine/render/voxels/palette.hpp>
 #include <cubos/engine/settings/plugin.hpp>
 #include <cubos/engine/settings/settings.hpp>
 #include <cubos/engine/transform/plugin.hpp>
@@ -22,8 +27,7 @@ int main(int argc, char** argv)
     cubos.plugin(windowPlugin);
     cubos.plugin(transformPlugin);
     cubos.plugin(assetsPlugin);
-    cubos.plugin(screenPickerPlugin);
-    cubos.plugin(rendererPlugin);
+    cubos.plugin(renderDefaultsPlugin);
 
     /// [Adding the plugin]
     cubos.plugin(voxelsPlugin);
@@ -33,14 +37,14 @@ int main(int argc, char** argv)
         settings.setString("assets.io.path", SAMPLE_ASSETS_FOLDER);
     });
 
-    cubos.startupSystem("create a camera").call([](Commands cmds, ActiveCameras& activeCameras) {
-        // Spawn the camera entity.
-        activeCameras.entities[0] = cmds.create()
-                                        .add(Camera{.fovY = 60.0F, .zNear = 0.1F, .zFar = 1000.0F})
-                                        .add(Position{{50.0F, 50.0F, 50.0F}})
-                                        .add(Rotation{glm::quatLookAt(glm::normalize(glm::vec3{-1.0F, -1.0F, -1.0F}),
-                                                                      glm::vec3{0.0F, 1.0F, 0.0F})})
-                                        .entity();
+    cubos.startupSystem("create a camera").call([](Commands cmds) {
+        auto targetEnt = cmds.create().add(RenderTargetDefaults{}).entity();
+
+        cmds.create()
+            .relatedTo(targetEnt, DrawsTo{})
+            .add(PerspectiveCamera{.fovY = 60.0F, .zNear = 0.1F, .zFar = 1000.0F})
+            .add(Position{{50.0F, 50.0F, 50.0F}})
+            .add(Rotation::lookingAt({-1.0F, -1.0F, -1.0F}, glm::vec3{0.0F, 1.0F, 0.0F}));
     });
 
     cubos.startupSystem("create the sun").call([](Commands cmds) {
@@ -49,15 +53,9 @@ int main(int argc, char** argv)
             .add(Rotation{glm::quat(glm::vec3(glm::radians(45.0F), glm::radians(45.0F), 0))});
     });
 
-    /// [Load and set palette]
-    cubos.startupSystem("load and set pallete")
-        .after(rendererInitTag)
-        .call([](const Assets& assets, Renderer& renderer) {
-            // Read the palette's data and pass it to the renderer.
-            auto palette = assets.read(PaletteAsset);
-            renderer->setPalette(*palette);
-        });
-    /// [Load and set palette]
+    /// [Set palette]
+    cubos.startupSystem("set palette").call([](RenderPalette& palette) { palette.asset = PaletteAsset; });
+    /// [Set palette]
 
     /// [Spawn car system]
     cubos.startupSystem("create a car").tagged(assetsTag).call([](Commands cmds, const Assets& assets) {
@@ -66,7 +64,7 @@ int main(int argc, char** argv)
         glm::vec3 offset = glm::vec3(car->size().x, 0.0F, car->size().z) / -2.0F;
 
         // Create the car entity
-        cmds.create().add(RenderableGrid{CarAsset, offset}).add(LocalToWorld{});
+        cmds.create().add(RenderVoxelGrid{CarAsset, offset}).add(LocalToWorld{});
     });
     /// [Spawn car system]